requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
